requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.7.14
charset-normalizer==3.4.2
distro==1.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
mutagen==1.47.0
openai-whisper==20231117
pydantic==2.11.7
pydantic_core==2.33.2
requests==2.32.4
sniffio==1.3.1
tqdm==4.67.1
twspace-dl==2024.7.2.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
